--- Jam Project/Assets/Player/Scripts/IPlayerAttackObserver.cs
@@ -1,6 +1,0 @@
-using UnityEngine;
-
-public interface IPlayerAttackObserver
-{
-    void NotifyAttack();
-}
