+++ Jam Project/Assets/Player/Scripts/PlayerAttackEffect.cs
@@ -1,35 +1,12 @@
 using System.Collections;
 using UnityEngine;
 
-public class PlayerAttackEffect : MonoBehaviour, IPlayerAttackObserver
+public class PlayerAttackEffect : PlayerAttackObserver
 {
     //private properties
     Animator _anim;
 
     //Private Methods
-
-    void OnEnable()
-    {
-        StartCoroutine(AddAttackObserver());
-    }
-
-    void OnDisable()
-    {
-        PlayerContext.Instance.RemoveAttackObserver(this);
-    }
-    IEnumerator AddAttackObserver()
-    {
-        for (; ; )
-        {
-            if (PlayerContext.Instance == null)
-            {
-                yield return null;
-            }
-            else break;
-        }
-        PlayerContext.Instance.AddAttackObserver(this);
-    }
-
 
     void Awake()
     {
@@ -40,7 +17,7 @@ public class PlayerAttackEffect : MonoBehaviour, IPlayerAttackObserver
 
     }
 
-    public void NotifyAttack()
+    public override void NotifyAttack()
     {
         _anim.SetTrigger("Attack");
     }
