+++ Jam Project/Assets/Player/Scripts/PlayerAttackObserver.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 698d78540cfb21943adb7bd1d04852c7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
