+++ Jam Project/Assets/Player/Scripts/PlayerCastAttack.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+public class PlayerCastAttack : PlayerAttackObserver
+{
+    public override void NotifyAttack()
+    {
+        Debug.Log("Casting Collider");
+    }
+}
