+++ Jam Project/Assets/Player/Scripts/PlayerCastAttack.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: abcb95f8c4935024da0768e359d92c1f
+guid: c50bc249e22f9fe449c66298a08f099e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
