+++ Jam Project/ProjectSettings/TagManager.asset
@@ -2,10 +2,11 @@
 %TAG !u! tag:unity3d.com,2011:
 --- !u!78 &1
 TagManager:
-  serializedVersion: 2
+  serializedVersion: 3
   tags:
   - BlackArea
   - WhiteArea
+  - Enemy_Land
   layers:
   - Default
   - TransparentFX
@@ -15,7 +16,7 @@ TagManager:
   - UI
   - Black
   - White
-  - 
+  - Enemy
   - 
   - 
   -
